+++ packages/terra-framework-docs/src/terra-dev-site/doc/worklist-data-grid/Examples.2/CellSelection.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useCallback } from 'react';
 import { DisclosureManagerContext } from 'terra-disclosure-manager';
 import WorklistDataGrid from 'terra-worklist-data-grid';
 import DisclosureComponent from './disclosure/DisclosureComponent';
@@ -59,38 +59,69 @@ const CellSelection = () => {
   const { cols, rows } = gridDataJSON;
   const disclosureManager = React.useContext(DisclosureManagerContext);
 
-  const onCellSelect = (rowId, columnId) => {
+  const [rowData, setRowData] = useState(rows);
+
+  const onCellSelect = useCallback((rowId, columnId) => {
     if (rowId && columnId) {
-      const rowIndex = rows.findIndex(e => e.id === rowId);
-      const colIndex = cols.findIndex(e => e.id === columnId);
+      const rowIndex = rowData.findIndex(e => e.id === rowId);
+      const columnIndex = cols.findIndex(e => e.id === columnId);
+
+      // Remove current selections
+      const newRowData = [...rowData];
+      for (let row = 0; row < rowData.length; row += 1) {
+        for (let cell = 0; cell < rowData[row].cells.length; cell += 1) {
+          const currentCell = rowData[row].cells[cell];
+          if (currentCell.isSelected && !(row === rowIndex && cell === columnIndex)) {
+            currentCell.isSelected = false;
+          }
+        }
+      }
+
+      // Toggle selection state of selected cell
+      newRowData[rowIndex].cells[columnIndex].isSelected = !rowData[rowIndex].cells[columnIndex].isSelected;
+      setRowData(newRowData);
+
       disclosureManager.disclose({
         preferredType: 'panel',
         size: 'tiny',
         content: {
           component: (
             <DisclosureComponent
-              columnHeader={cols[colIndex].displayName}
-              rowHeader={rows[rowIndex].cells[0].content}
-              content={rows[rowIndex].cells[colIndex].content}
-              name={cols[colIndex].displayName}
+              columnHeader={cols[columnIndex].displayName}
+              rowHeader={rowData[rowIndex].cells[0].content}
+              content={rowData[rowIndex].cells[columnIndex].content}
+              name={cols[columnIndex].displayName}
             />
           ),
         },
       });
     }
-  };
+  }, [cols, disclosureManager, rowData]);
+
+  const onClearSelectedCells = useCallback(() => {
+    // Remove current selections
+    const newRowData = [...rowData];
+    for (let row = 0; row < rowData.length; row += 1) {
+      for (let cell = 0; cell < rowData[row].cells.length; cell += 1) {
+        newRowData[row].cells[cell].isSelected = false;
+      }
+    }
+
+    setRowData(newRowData);
+  }, [rowData]);
 
   return (
     <WorklistDataGrid
       id="worklist-data-grid-row-selection"
       overflowColumns={cols}
-      rows={[...rows]}
+      rows={rowData}
       rowHeaderIndex={rowHeaderIndex}
       rowHeight="50px"
       defaultColumnWidth={100}
       columnHeaderHeight="100px"
       ariaLabel="Worklist Data Grid With Cell Selection"
       onCellSelect={onCellSelect}
+      onClearSelectedCells={onClearSelectedCells}
     />
   );
 };