+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Changed
   * Updated focusable cell example for `terra-worklist-data-grid` to allow text area to be editable.
+  * Updated examples and tests for `terra-worklist-data-grid` to account for cell selection state being controlled by consumers.
 
 ## 1.32.0 - (August 10, 2023)